+++ Runtime/Scripts/Definitions/ASoundEffectDefinition.cs
@@ -4,6 +4,6 @@ namespace JakubKrizanovsky.SimpleSoundSystem
 {
 	public abstract class ASoundEffectDefinition : ScriptableObject
 	{
-		public abstract void InitializeSoundEffect(SoundEffect soundEffect, float pitchMultiplier);
+		internal abstract void InitializeSoundEffect(SoundEffect soundEffect, SoundPlayParameters parameters);
 	}
 }