+++ Runtime/Scripts/SoundPlayParameters.cs
@@ -1,0 +1,11 @@
+using System;
+
+namespace JakubKrizanovsky.SimpleSoundSystem
+{
+    [Serializable]
+    public class SoundPlayParameters
+    {
+        public float VolumeMultiplier = 1f;
+        public float PitchMultiplier = 1f;
+    }
+}