+++ CHANGELOG.md
@@ -3,3 +3,9 @@
 - Sound effect player singleton
 - Sound effect script
 - Two simple sound effect definitions (scriptable objects)
+
+## [1.1.0] - 2025-11-16
+### Volume Support
+- Added support for controlling volume of played sounds
+- Reworked API for simpler usage (breaking API changes)
+- Custom editor to test sound effect definitions